+++ app/models/component/manifest.rb
@@ -31,10 +31,12 @@ class Component
 
     def spec_template
       if Configurations.apps.node_selector.nil?
-        { metadata:     { labels: labels },
+        { metadata:     { labels: labels,
+                          annotations: annotations },
           spec:         { containers: containers } }.with_indifferent_access
       else
-        { metadata:     { labels: labels },
+        { metadata:     { labels: labels,
+                          annotations: annotations },
           spec:         { containers: containers,
                           nodeSelector: { "beta.kubernetes.io/instance-type": \
                                           Configurations.apps.node_selector \
@@ -47,6 +49,10 @@ class Component
         "#{prefix}/environment" => environment.name,
         "#{prefix}/component" => type,
         "#{prefix}/nodetype" => node.id }.with_indifferent_access
+    end
+
+    def annotations
+      { "ad.datadoghq.com/#{name}.logs": '[{"source":"%%env_DATADOG_LOGS_SOURCE%%","service":"%%env_DATADOG_LOGS_SERVICE%%"}]'}.with_indifferent_access
     end
 
     def spec